Move the smoke testing to one place.
I'm not sure about using diagnostics, we probably
need our own objects, so this is going to grow even a bit
more, may as well put it in one place.
